pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Aleksandr Platcev <a.platcev@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.0"
django-environ = "^0.11.2"
psycopg = {extras = ["binary"], version = "^3.1.12"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
